Placeholder creations and class migration
+Created Placeholders for each class
+Moved existing classes to corresponding placeholders
