+++ ProductManagementSystem/src/main/java/com/dev/productmanagementsystem/!TODO
@@ -1,10 +1,10 @@
 AFTER TASK IS DONE PUT A "+" INSTEAD OF "-"\
 TODO:
--Create Placeholders for each class
--Move existing classes to corresponding placeholders
++Create Placeholders for each class
++Move existing classes to corresponding placeholders
 -Refactor the code so it doesn't create any errors
 -update dependencies
--update application.properties
+?update application.properties (may have an issue with critical data so currently ignore this)
 -Verify the code whether it has everything that has been declared in the class diagram
 -Check what's going on with these enums(and whether are they necessary)
 -Make necessary updates to the classes creating all the methods that are needed according to the diagram
