+++ ProductManagementSystem/src/main/java/com/dev/productmanagementsystem/OrderStatus.java
@@ -1,0 +1,9 @@
+package com.dev.productmanagementsystem;
+
+public enum OrderStatus {
+    PENDING,
+    PROCESSING,
+    SHIPPED,
+    DELIVERED,
+    CANCELLED
+}
