+++ ProductManagementSystem/src/main/java/com/dev/productmanagementsystem/PaymentMethod.java
@@ -1,0 +1,9 @@
+package com.dev.productmanagementsystem;
+
+public enum PaymentMethod {
+    CREDIT_CARD,
+    DEBIT_CARD,
+    PAYPAL,
+    BANK_TRANSFER,
+    CASH
+}
