+++ ProductManagementSystem/src/main/java/com/dev/productmanagementsystem/PaymentStatus.java
@@ -1,0 +1,9 @@
+package com.dev.productmanagementsystem;
+
+public enum PaymentStatus {
+    PENDING,
+    PAID,
+    FAILED,
+    REFUNDED,
+    CANCELLED
+}
