+++ ProductManagementSystem/src/main/java/com/dev/productmanagementsystem/Permission.java
@@ -1,0 +1,32 @@
+package com.dev.productmanagementsystem;
+
+import java.util.Set;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "permissions")
+public class Permission {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String name;
+
+    @ManyToMany(mappedBy = "permissions")
+    private Set<Role> roles;
+
+    // Constructors
+    public Permission() {}
+
+    // Getters and Setters
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+
+    public Set<Role> getRoles() { return roles; }
+    public void setRoles(Set<Role> roles) { this.roles = roles; }
+}
+
