--- ProductManagementSystem/src/main/java/com/dev/productmanagementsystem/enums.java
@@ -1,29 +1,0 @@
-package com.dev.productmanagementsystem;
-
-public enum OrderStatus {
-    PENDING,
-    PROCESSING,
-    SHIPPED,
-    DELIVERED,
-    CANCELLED
-}
-
-package com.dev.productmanagementsystem;
-
-public enum PaymentStatus {
-    PENDING,
-    PAID,
-    FAILED,
-    REFUNDED,
-    CANCELLED
-}
-
-package com.dev.productmanagementsystem;
-
-public enum PaymentMethod {
-    CREDIT_CARD,
-    DEBIT_CARD,
-    PAYPAL,
-    BANK_TRANSFER,
-    CASH
-}
